# [core] reduce unnecessary traversals to optimize the assign logic of PartitionIndex.