--- a/paimon-core/src/main/java/org/apache/paimon/index/HashBucketAssigner.java
+++ b/paimon-core/src/main/java/org/apache/paimon/index/HashBucketAssigner.java
@@ -157,6 +157,7 @@ public class HashBucketAssigner {
                 indexFileHandler,
                 partition,
                 targetBucketRowNumber,
-                (hash) -> computeAssignId(hash) == assignId);
+                (hash) -> computeAssignId(hash) == assignId,
+                (bucket) -> computeAssignId(bucket) == assignId);
     }
 }